+++ NEWS.md
@@ -14,6 +14,7 @@
 - Prevent attempting to start Copilot on a non-main thread (#14952)
 - Reformat Code no longer inserts whitespace around '^' operator (#14973)
 - Prompt for personal access token instead of password when using github via https (#14103)
+- RStudio now forward the current 'repos' option for actions taken in the Build pane (#5793)
 
 #### Posit Workbench
 -